backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
sqlalchemy==2.0.31
chromadb==0.5.3
google-generativeai==0.7.2
python-dotenv==1.0.1
pytest==8.2.2
httpx==0.27.0
scikit-learn==1.5.0
pandas==2.2.2
